+++ src/index.js
@@ -9,16 +9,18 @@ if (!observable) throw new Error("mobx-react requires mobx to be available")
 if (typeof rdBatched === "function") configure({ reactionScheduler: rdBatched })
 else if (typeof rnBatched === "function") configure({ reactionScheduler: rnBatched })
 
-// TODO: re-export more mobx-react-lite stuff?
-// TODO; do we still need separate RN build?
-export { Observer } from "mobx-react-lite"
+export {
+    useObservable,
+    useComputed,
+    useDisposable,
+    IObserverOptions,
+    useObserver,
+    Observer
+} from "mobx-react-lite"
 
 export { observer, useStaticRendering } from "./observer"
 
-export { default as Provider } from "./Provider"
-export { default as inject } from "./inject"
+export { Provider } from "./Provider"
+export { inject } from "./inject"
 export { disposeOnUnmount } from "./disposeOnUnmount"
-
-import * as propTypes from "./propTypes"
-export { propTypes }
-export { propTypes as PropTypes }
+export { PropTypes } from "./propTypes"